erdjs min, for browser.
